Add `depthconv_conv2d` tests
